% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_graph.R
\name{add_value_to_labels}
\alias{add_value_to_labels}
\title{Add value labels for the treemap graph}
\usage{
add_value_to_labels(data, with_labels)
}
\arguments{
\item{data}{Tibble.}

\item{with_labels}{Logical.}
}
\value{
tibble with category labels with values added
}
\description{
Add value labels for the treemap graph
}
\author{
Leonardo Rocha
}
