% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI_survey.R
\docType{data}
\name{AI_survey}
\alias{AI_survey}
\title{AI Survey dataset}
\format{
a list with tree tibbles:
\describe{
\item{survey_data}{tibble with  31 questions and 1430 answers from AI survey}
\item{answers_labels}{tibble with answers labels for all answer options in survey data}
\item{questions_labels}{tibble with questions labels for all the 31 questions of the AI survey dataset}
}
}
\usage{
AI_survey
}
\description{
AI Survey dataset with question and answers labels
}
\keyword{datasets}
