% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_aux.R
\name{apply_labels}
\alias{apply_labels}
\title{Apply Answers and Questions Labels}
\usage{
apply_labels(survey_data, question_label, answer_label)
}
\arguments{
\item{survey_data}{Tibble. Dados já organizados com os percentuais calculados}

\item{question_label}{Tibble. Dados gerados pelo usuário com os rótulos
de questão vinculado aos identificadores de questão}

\item{answer_label}{Tibble. Dados gerados pelo usuário com os rótulos
de respostas vinculadas aos identificadores de questão}
}
\value{
Tibble. The survey_data with
}
\description{
Apply answers and questions labels to the survey dataset.
}
\author{
Leonardo Rocha
}
