% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_data.R
\name{apply_labels_answer}
\alias{apply_labels_answer}
\title{Apply answers label to survey data}
\usage{
apply_labels_answer(data, answer)
}
\arguments{
\item{data}{Tibble.}

\item{answer}{Tibble.}
}
\value{
Tibble.
}
\description{
Apply answers label to survey data
}
\author{
Leonardo Rocha
}
