% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_text.R
\name{break_text}
\alias{break_text}
\title{Break text}
\usage{
break_text(string, max_size)
}
\arguments{
\item{string}{Character. Text to break}

\item{max_size}{Numeric. Max size line parameter}
}
\value{
Character. Text containing one or more lines
}
\description{
Break a text into multiple lines
}
\author{
Leonardo Rocha
}
