% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_aux.R
\name{chi.square_pvalue}
\alias{chi.square_pvalue}
\title{Compute Chi-square p-value}
\usage{
chi.square_pvalue(contingency.table)
}
\arguments{
\item{contingency.table}{Tibble. With the contingency table between two categorical questions}
}
\value{
Numeric. Chi-square p-value
}
\description{
\code{chi.square_pvalue} compute the Chi-squared test, with simulated p-values and get its p-value.
}
\details{
The function is applied to a contingency table between two categorical variables, and
return the p-value of the Chi-square test though the usage of the \code{stats::chisq.test}
function.
}
\author{
Leonardo Rocha
}
