% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{filter_data}
\alias{filter_data}
\title{Filter other and missing data}
\usage{
filter_data(data, rm.other, other_label = "Other")
}
\arguments{
\item{data}{Tibble.}

\item{rm.other}{Logical.}

\item{other_label}{Character.}
}
\value{
Tibble.
}
\description{
Filter other and missing data
}
\author{
Leonardo Rocha
}
