% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph_aux.R
\name{gen_axis_y_scale}
\alias{gen_axis_y_scale}
\title{Gerar rótulos do eixo y}
\usage{
gen_axis_y_scale(data, negative_col, intervals = 0.25)
}
\arguments{
\item{data}{Tibble.}

\item{negative_col}{Character. Nome da coluna que é negativa}

\item{intervals}{Numeric. Tamanho do intervalo da escala de valores continuos do eixo y}
}
\value{
Escala de valores para o eixo y do gráfico de barras
}
\description{
Gerar rótulos do eixo y
}
