% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph_aux.R
\name{gen_bar_label}
\alias{gen_bar_label}
\title{add label}
\usage{
gen_bar_label(
  tidy_data,
  with_labels,
  label_thereshold = 0.05,
  percent_accuracy = 0.01,
  alignment = 0.5,
  text_size = 2
)
}
\arguments{
\item{tidy_data}{Tibble. Dados da questão principal organizados para a elaboração do gráfico de barras.}

\item{with_labels}{Logical. Indicador de se o gráfico apresenta rótulo de texto.}

\item{label_thereshold}{Numeric. Limiar do percentual para que seja atribuido um rotulo à observação.}

\item{percent_accuracy}{Numeric. Precisão das casas decimais do percentual.}

\item{alignment}{Numeric. Alinhamento do rótulo de texto.}

\item{text_size}{Numeric. Tamanho dos rótulos de texto.}
}
\description{
Adiciona rótulos ao gráfico
}
