% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph_aux.R
\name{gen_fill_scale}
\alias{gen_fill_scale}
\title{Gera a escala de cores}
\usage{
gen_fill_scale(
  tidy_data,
  n_negative,
  intermediate,
  diverg_scale,
  negative_palette,
  positive_palette,
  sequential_palette
)
}
\arguments{
\item{tidy_data}{Tibble. Dados arrumados e prontos para serem utilizados na geração dos gráficos}

\item{n_negative}{Numeric. Número de escalas negativas}

\item{intermediate}{Logic. Existe escala intermediária}

\item{diverg_scale}{Logic. Se a escala de cores vai seguir uma lógica divergente}

\item{negative_palette}{Character. O nome da paleta de cores para a escala negativa}

\item{positive_palette}{Character. O nome da paleta de cores para a escala positiva}

\item{sequential_palette}{Character. O nome da paleta de cores para a escala sequencial}
}
\value{
ggplot2::scale_fill_manual com a escala para preenchimento do gráfico de barras
}
\description{
Gera a escala de cores para o gráfico de barras
}
\author{
Leonardo Rocha
}
