% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph_aux.R
\name{gen_graph_theme}
\alias{gen_graph_theme}
\title{Gerar removedor de rótulos do eixo x}
\usage{
gen_graph_theme(base_text_size, legend_position, data)
}
\arguments{
\item{base_text_size}{Numeric. Tamanho base dos textos para o tema.}

\item{legend_position}{Character. Indicativo da posição da legenda.}

\item{data}{Tibble. Dados com a tabulação da questão principal.}
}
\description{
Gerar removedor de rótulos do eixo x
}
