% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_sub_question_data}
\alias{gen_sub_question_data}
\title{Generate sub_question data}
\usage{
gen_sub_question_data(sub, n_answer, min_count, max_count)
}
\arguments{
\item{sub}{Numeric. The index of the sub_question that is being generated}

\item{n_answer}{Numeric. The number of answer options}

\item{min_count}{Numeric. Minimum count value that an answer will receive}

\item{max_count}{Numeric. Maximum count value that an answer will receive}
}
\value{
Sub-question answer stats data
}
\description{
Helper function that generates sub_questions answers stats
for the Generate Random Question function. The answers count is
generated through a uniform probability distribution.
}
\author{
Leonardo Rocha
}
