% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph_aux.R
\name{gen_title}
\alias{gen_title}
\title{Gerador de título}
\usage{
gen_title(question_data, title_max_size, with_title)
}
\arguments{
\item{question_data}{Tibble. Dados organizados relativos a uma questão principal}

\item{title_max_size}{Numeric. Tamanho máximo da linha do título}

\item{with_title}{Logical. Título deve ser gerado ou não}
}
\value{
Character. Título a ser utilizado no gráfico
}
\description{
Gerador do título para o gráfico de barras
}
\author{
Leonardo Rocha
}
