% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph_aux.R
\name{get_lower_limit}
\alias{get_lower_limit}
\title{Get axis lower limit}
\usage{
get_lower_limit(data, negative_col, intervals, gap = 1.1)
}
\arguments{
\item{data}{Tibble}

\item{negative_col}{Character.}

\item{intervals}{Numeric.}

\item{gap}{Numeric.}
}
\value{
Numeric.
}
\description{
Get axis lower limit for the negative side of a likert scale graph with a negative connotation answer.
}
\author{
Leonardo Rocha
}
