% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_aux.R
\name{get_n_questions}
\alias{get_n_questions}
\title{Get number of questions}
\usage{
get_n_questions(survey_data)
}
\arguments{
\item{survey_data}{Tibble. Question table with numbers as categories}
}
\value{
Numeric. The number of questions containing in the survey data
}
\description{
Get the number of questions from the survey data.
}
\author{
Leonardo Rocha
}
