% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(file_path)
}
\arguments{
\item{file_path}{Character. path to the XLSX file}
}
\value{
List. List with tree named Tibbles, 'data', 'questions' and 'answers'.
}
\description{
Load the survey data from an excel file. The data in the excel must be in a very specific form.
}
\author{
Leonardo Rocha
}
