% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph.R
\name{make_bar_graph}
\alias{make_bar_graph}
\title{Make a bar graph}
\usage{
make_bar_graph(
  question_data,
  negative_col = FALSE,
  fill_color = "Oranges",
  title_max_size = 60,
  legend_max_size = 10,
  axis_max_size = 25,
  legend_position = "bottom",
  base_text_size = 9,
  has_title = FALSE,
  with_labels = TRUE
)
}
\arguments{
\item{question_data}{Tibble. Question stats table}

\item{negative_col}{Character. Name of the category with negative connotation, if \code{FALSE}
the answers will be ordered by their prevalence. Default to \code{FALSE}.}

\item{fill_color}{Character. Color scale name from ggplot2 library. Default \code{'Oranges'}.}

\item{title_max_size}{Numeric. Max line size for the title text. Default \code{60}.}

\item{legend_max_size}{Numeric. Max line size for the legend text. Default \code{10}.}

\item{axis_max_size}{Numeric. Max line size for the x axis text. Default \code{25}.}

\item{legend_position}{Character. \code{ggplot2} \code{legend_position} character indicator. Default \code{"bottom"}.}

\item{base_text_size}{Numeric. \code{ggplot2} \code{theme} \code{base_size} for all the text elements. Default \code{9}.}

\item{has_title}{Logical. Indicative if the graph will have a title. The main question label will be used for the
title text. Default \code{FALSE}.}

\item{with_labels}{Logical.Indicative if the graph will have text indicating the values inside the bar graph.
To make the figure cleaner answers with prevalence bellow 5\% won't have text. Default \code{TRUE}}
}
\value{
Custom horizontal bar graph: a likert-scale or a ordered by value answers.
}
\description{
Function to make a likert scale bar graph or ordered by values bar graph
}
\details{
Make horizontal bar graph with the \code{question_data} provided.
If a \code{negative_col} is provided it will be a likert-scale graph with
the \code{negative_col} been the answer with a negative connotation. Else,
the answers categories in the bar graph will be ordered according to their
prevalence.
}
\author{
Leonardo Rocha
}
