% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_graph.R
\name{make_combined_treemap_graph}
\alias{make_combined_treemap_graph}
\title{Make a combined treemap graph}
\usage{
make_combined_treemap_graph(
  grouped_data,
  with_labels = TRUE,
  fill_color = "Set2",
  legend_max_size = 15,
  base_text_size = 9,
  legend_position = "bottom",
  treemap_proportion = 0.9
)
}
\arguments{
\item{grouped_data}{Tibble.}

\item{with_labels}{Logical. Default \code{TRUE}}

\item{fill_color}{Character. Default \code{"Set2"}}

\item{legend_max_size}{Numeric. Default \code{15}}

\item{base_text_size}{Numeric. Default \code{9}}

\item{legend_position}{Character. Default \code{"bottom"}}

\item{treemap_proportion}{Numeric. Default \code{0.9}}
}
\value{
A ggplot2 figure with bar and treemap graphs combined
}
\description{
Similar to \code{make_treemap_graph} this function takes a grouped questions tibble
and produces a treemap graph. Along with a treemap it is also genereted a bar graph,
with the distribution of the main group. Both graphs are vertically combined into a single figure,
with the treemap graph taking \code{treemap_proportion} upper part.
}
\author{
Leonardo Rocha
}
