% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_graph.R
\name{make_complementar_bar_graph}
\alias{make_complementar_bar_graph}
\title{Make a complementar bar graph}
\usage{
make_complementar_bar_graph(
  grouped_data,
  with_labels = TRUE,
  fill_color = "Set2",
  legend_max_size = 15,
  base_text_size = 9,
  legend_position = "none"
)
}
\arguments{
\item{grouped_data}{Tibble.}

\item{with_labels}{Logical. Default \code{TRUE}}

\item{fill_color}{Character. Default \code{"Set2"}}

\item{legend_max_size}{Numeric. Default \code{15}}

\item{base_text_size}{Numeric. Default \code{9}}

\item{legend_position}{Character. Default \code{"none"}}
}
\value{
A ggplot2 bar graph
}
\description{
Make a complementar bar graph for the treemap combined graph
}
\author{
Leonardo Rocha
}
