% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_graph.R
\name{make_treemap_graph}
\alias{make_treemap_graph}
\title{Make treemap graph}
\usage{
make_treemap_graph(
  grouped_data,
  with_labels = TRUE,
  fill_color = "Set2",
  legend_max_size = 15,
  base_text_size = 9,
  legend_position = "bottom"
)
}
\arguments{
\item{grouped_data}{Tibble.}

\item{with_labels}{Logical. Default \code{TRUE}}

\item{fill_color}{Character. Default \code{"Set2"}}

\item{legend_max_size}{Numeric. Default \code{15}}

\item{base_text_size}{Numeric. Default \code{9}}

\item{legend_position}{Character. Default \code{"bottom"}}
}
\value{
A ggplot2 treemap graph
}
\description{
Function that takes grouped questions data and make a treemap graph
}
\author{
Leonardo Rocha
}
