% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{make_wilcox_test}
\alias{make_wilcox_test}
\title{Make a Wilcoxon rank sum test}
\usage{
make_wilcox_test(surveydata, qnumber)
}
\arguments{
\item{surveydata}{Tibble.}

\item{qnumber}{Numeric.}
}
\value{
Tibble with wicolxon stats
}
\description{
Make a Wilcoxon rank sum test between answer and knowledge
}
\author{
Leonardo Rocha
}
