% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tab_all_survey_question}
\alias{tab_all_survey_question}
\title{Tabulate all survey questions}
\usage{
tab_all_survey_question(surveydata)
}
\arguments{
\item{surveydata}{List. A list of composed of tibbles of survey data, answers labels and questions labels, generated by \code{load_data} function.}
}
\value{
List. A list of Tables. A Table for each question
with percent values and p-value (chi-squared test between knowledge level)
for sub-questions and knowledge levels.
}
\description{
Apply \code{tab_survey_question} to all questions in a survey data
}
\author{
Leonardo Rocha
}
