% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{tab_grouped_questions}
\alias{tab_grouped_questions}
\title{Tabulate grouped questions}
\usage{
tab_grouped_questions(
  surveydata,
  group1_col,
  group2_col,
  rm.other = TRUE,
  unite.sep = "__"
)
}
\arguments{
\item{surveydata}{List.}

\item{group1_col}{Character.}

\item{group2_col}{Vector.}

\item{rm.other}{Logical.}

\item{unite.sep}{Character.}
}
\value{
Tibble.
}
\description{
Generate a table from questions that have some dependency relationship
}
\details{
The \code{tab_grouped_questions} takes a \code{surveydata}, a main independent question (\code{group1_col}) and a list of
main questions (\code{group2_col}) that are dependent on the main independent question and generate a table that relates
answers in both group of questions.
}
\author{
Leonardo Rocha
}
