% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tab_survey_question}
\alias{tab_survey_question}
\title{Tabulate a survey question}
\usage{
tab_survey_question(surveydata, question_number)
}
\arguments{
\item{surveydata}{List. A list of composed of tibbles of survey data, answers labels and questions labels, generated by \code{load_data} function.}

\item{question_number}{Numeric. Number of the main question}
}
\value{
Tibble. Table with percent values and p-value (chi-squared test between knowledge level) for sub-questions and knowledge levels.
}
\description{
\code{tab_survey_question} Create a statistical table for a target question
}
\details{
This function create a tibble table, with percent values for each answer and knowledge level for
all sub-question (if any) regarding the target main question. It takes a survey data, select the main question,
with it's respectives sub-questions, and calculate percent values for each answer, sub-question and knowledge level.
Also for each sub-question, it makes a chi-square test for the correlation between responses and knowledge level.
Finally, the function applies the question and answers labels.
}
\examples{
tab_survey_question(surveydata = AI_survey, question_number = 4)
tab_survey_question(AI_survey, 2)
}
\author{
Leonardo Rocha
}
