% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph_aux.R
\name{tidy_data_for_bar_graph}
\alias{tidy_data_for_bar_graph}
\title{Arrumar dados para gráfico de barra}
\usage{
tidy_data_for_bar_graph(
  question_data,
  n_negative,
  intermediate,
  legend_max_size,
  axis_max_size
)
}
\arguments{
\item{question_data}{Tibble. Dados organizados relativos a uma questão principal}

\item{n_negative}{Numeric. Número de escalas negativas}

\item{intermediate}{Logic. Existe escala intermediária}

\item{legend_max_size}{Numeric. Tamanho máximo da linha do texto da legenda do gráfico}

\item{axis_max_size}{Numeric. Tamanho máximo da linha do texto do eixo x do gráfico}
}
\value{
Tibble. Dados arrumados e prontos para serem utilizados na geração dos gráficos
}
\description{
Arrumar os dados para o gráfico de barras
}
\details{
Preparação dos dados para a geração dos gráficos
Arruma o dados de acordo com a escala das respostas:
se existem escalas negativas e se existem escalas intermediárias
a regra da escala intermediária só será aplicada caso haja alguma escala negativa
pressupondo que as escalas estão em ordem crescente para indicar se há escalas
negativas, basta informar a quantidade de respostas com conotação negativa
se houver escala intermediária era será equivalente à resposta seguinte às
escalas negativas. Caso não haja escala negativa, a escala de respostas será
ordenada por suas frequências relativa. Caso haja escala negativa, os de valores das
frequências relativa das escalas negativas serão transformados para o valor negativo.
Caso haja escala negativa e intermediária, além dos valores negativos serem transformados
para o negativo, o valor da escala intermediária será dividido por dois e duplicado
sendo um mantendo o mesmo sinal e o outro assumindo o sinal negativo.
}
\author{
Leonardo Rocha
}
