% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{wilcox_ind_sample}
\alias{wilcox_ind_sample}
\title{Utility function for the Wicoxon rank sum test}
\usage{
wilcox_ind_sample(nested_data)
}
\description{
@param nested_data Tibble.
}
\details{
@return row with wilcoxcon rank sum test for two idependent

@author Leonardo Rocha
}
