% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_aux.R
\name{apply_labels}
\alias{apply_labels}
\title{Apply Answers and Questions Labels}
\usage{
apply_labels(survey_data, question_label, answer_label)
}
\arguments{
\item{survey_data}{Tibble. Question table with numbers as categories}

\item{question_label}{Tibble. Table of question labels}

\item{answer_label}{Tibble. Table of answer labels}
}
\value{
Tibble. The survey_data with
}
\description{
Apply answers and questions labels to the survey dataset.
}
\author{
Leonardo Rocha
}
