% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph_aux.R
\name{gen_bar_label}
\alias{gen_bar_label}
\title{Rótulo para gráfico de barras}
\usage{
gen_bar_label(
  question_data,
  tidy_data,
  n_negative,
  label_thereshold,
  label_percent_accuracy,
  label_text_size,
  with_labels
)
}
\arguments{
\item{question_data}{Tibble. Dados organizados relativos a uma questão principal}

\item{tidy_data}{Tibble. Dados arrumados e prontos para serem utilizados na geração dos gráficos}

\item{n_negative}{Numeric. Número de escalas negativas}

\item{label_thereshold}{Numeric. Limiar do percentual para que seja atribuido um rotulo à observação.}

\item{label_percent_accuracy}{Numeric. Precisão das casas decimais do percentual.}

\item{label_text_size}{Numeric. Tamanho dos rótulos de texto.}

\item{with_labels}{Logical. Indicador de se o gráfico apresenta rótulo de texto.}
}
\value{
ggplot2::geom_text com os rótulos para o gráfico de barras (caso with_labels)
}
\description{
Gerado de rótulos para o gráfico de barras
}
\author{
Leonardo Rocha
}
