% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph_aux.R
\name{gen_fill_scale}
\alias{gen_fill_scale}
\title{Gerar escala de cores}
\usage{
gen_fill_scale(
  palette_color,
  direction_color = 1,
  grey_start = 0.7,
  grey_end = 0.3
)
}
\arguments{
\item{palette_color}{Character. Nome da paleta de cores utilizadas caso não seja em escala de cinza}

\item{direction_color}{Numeric. Direção da paleta de cores}

\item{grey_start}{Numeric. Intensidade de cinza do começo da escala de cinza}

\item{grey_end}{Numeric. Intensidade de cinza do fim da escala de cinza}
}
\value{
Escala de cor para o gráfico de barras
}
\description{
Gerar escala de cores para o gráfico de barras
}
\author{
Leonardo Rocha
}
