% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph_aux.R
\name{gen_graph_theme}
\alias{gen_graph_theme}
\title{Gerador de tema de gráfico em barras}
\usage{
gen_graph_theme(base_text_size, legend_position, question_data)
}
\arguments{
\item{base_text_size}{Numeric. Tamanho base dos textos para o tema.}

\item{legend_position}{Character. Indicativo da posição da legenda.}

\item{question_data}{Tibble. Dados com a tabulação da questão principal.}
}
\value{
ggplot2::theme com a configuração do tema para o gráfico de barras
}
\description{
Gera o tema para ser utilizado no gráfico de barras.
}
\author{
Leonardo Rocha
}
