% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph.R
\name{make_bar_graph}
\alias{make_bar_graph}
\title{Geração de gráfico de barras}
\usage{
make_bar_graph(
  question_data,
  n_negative,
  intermediate,
  diverg_scale = F,
  negative_palette = "Reds",
  positive_palette = "Greens",
  sequential_palette = "Oranges",
  title_max_size = 60,
  with_title = FALSE,
  legend_max_size = 10,
  axis_max_size = 25,
  axis_y_interval = 0.25,
  axis_y_fixed = TRUE,
  axis_y_gap = 1.1,
  legend_position = "bottom",
  base_text_size = 9,
  with_labels = TRUE,
  label_thereshold = 0.05,
  label_percent_accuracy = 0.2,
  label_text_size = 2
)
}
\arguments{
\item{question_data}{Tibble. Dados organizados relativos a uma questão principal}

\item{n_negative}{Numeric. Número de escalas negativas}

\item{intermediate}{Logic. Existe escala intermediária}

\item{diverg_scale}{Logic. Se a escala de cores vai seguir uma lógica divergente}

\item{negative_palette}{Character. O nome da paleta de cores para a escala negativa}

\item{positive_palette}{Character. O nome da paleta de cores para a escala positiva}

\item{sequential_palette}{Character. O nome da paleta de cores para a escala seq}

\item{title_max_size}{Numeric. Tamanho máximo da linha do título}

\item{with_title}{Logical. Título deve ser gerado ou não}

\item{legend_max_size}{Numeric. Tamanho máximo da linha do texto da legenda do gráfico}

\item{axis_max_size}{Numeric. Tamanho máximo da linha do texto do eixo x do gráfico}

\item{axis_y_interval}{Numeric. Tamanho do intervalo entre os rótulos do eixo}

\item{axis_y_fixed}{Logical. Se o eixo vai ser fixo ou ajustado aos dados}

\item{axis_y_gap}{Numeric. Tamanho da margem para o ajuste aos dados}

\item{legend_position}{Character. Indicativo da posição da legenda.}

\item{base_text_size}{Numeric. Tamanho base dos textos para o tema.}

\item{with_labels}{Logical. Indicador de se o gráfico apresenta rótulo de texto.}

\item{label_thereshold}{Numeric. Limiar do percentual para que seja atribuido um rotulo à observação.}

\item{label_percent_accuracy}{Numeric. Precisão das casas decimais do percentual.}

\item{label_text_size}{Numeric. Tamanho dos rótulos de texto.}
}
\value{
ggplot2::ggplot de gráfico de barras
}
\description{
Função para a geração de gráfico de barras para questões de survey categóricas.
}
\details{
Essa função pode ser utilizada para a criação de gráficos em barra,
podendo gerar gráficos de barras divergentes ou sequenciais empilhadas.
Múltiplas configurações podem ser realizadas através da modificação dos
pâmetros. Como por exemplo, adicionar ou remover títulos e rótulos aos
dados, modificar o tamanho dos textos, modifcar as cores a serem utilizadas
nas escalas divergentes e sequenciais etc.
}
\author{
Leonardo Rocha
}
