% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveystats.R
\docType{package}
\name{surveystats}
\alias{surveystats}
\alias{surveystats-package}
\title{Survey Stats Analysis}
\description{
This package has functionalities regarding the statistical and graphical analysis
of Survey Monkey survey research data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leofmr/surveystats}
  \item Report bugs at \url{https://github.com/leofmr/surveystats/issues}
}

}
\author{
Leonardo Rocha \email{leonardo.rocha@fiocruz.br}
}
