% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tab_all_survey_question}
\alias{tab_all_survey_question}
\title{Tabulate all survey questions}
\usage{
tab_all_survey_question(survey_data, question_label, answer_label)
}
\arguments{
\item{survey_data}{Tibble. Survey data}

\item{question_label}{Tibble. Question labels data}

\item{answer_label}{Tibble. Answers labels data}
}
\value{
List. A list of Tables. A Table for each question
with percent values and p-value (chi-squared test between knowledge level)
for sub-questions and knowledge levels.
}
\description{
Tabulate all survey questions
}
\author{
Leonardo Rocha
}
