% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tab_survey_question}
\alias{tab_survey_question}
\title{Tabulate a survey question}
\usage{
tab_survey_question(survey_data, question_number, question_label, answer_label)
}
\arguments{
\item{survey_data}{Tibble. Survey data}

\item{question_number}{Numeric. Number of the main question}

\item{question_label}{Tibble. Question labels data}

\item{answer_label}{Tibble. Answers labels data}
}
\value{
Tibble. Table with percent values and p-value (chi-squared test between knowledge level) for sub-questions and knowledge levels.
}
\description{
\code{tab_survey_question} Create a statistical table for a target question
}
\details{
This function create a tibble table, with percent values for each answer and knowledge level for
all sub-question (if any) regarding the target main question. It takes a survey data, select the main question,
with it's respectives sub-questions, and calculate percent values for each answer, sub-question and knowledge level.
Also for each sub-question, it makes a chi-square test for the correlation between responses and knowledge level.
Finally, the function applies the question and answers labels.
}
\examples{
tab_survey_question(AI_survey$survey_data, 4, AI_survey$questions_labels, AI_survey$answers_labels)
tab_survey_question(AI_survey$survey_data, 2, AI_survey$questions_labels, AI_survey$answers_labels)
}
\author{
Leonardo Rocha
}
