% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph_aux.R
\name{tidy_data_for_bar_graph}
\alias{tidy_data_for_bar_graph}
\title{Arrumar dados para gráfico de barra}
\usage{
tidy_data_for_bar_graph(data, negative_col, legend_max_size, axis_max_size)
}
\arguments{
\item{data}{Tibble. survey dataset completo}

\item{negative_col}{Character. Nome da questão que possui a conotação negativa}

\item{legend_max_size}{Numeric. Tamanho máximo da linha do texto legenda do gráfico}

\item{axis_max_size}{Numeric. Tamanho máximo da linha do texto do eixo x do gráfico}
}
\value{
Tibble. Dados arrumados e prontos para serem utilizados na geração dos gráficos
}
\description{
Função para arrumar os gráficos para o gráfico de barras
}
\details{
Preparação dos dados para a geração do gráfico
Apenas todos os conhecimentos serão selecionados
os valores percentuais das linhas com a escolha de conotação negativa
são transformados para o negativo
foi feito de uma certa forma que, caso negative_col seja equivalente a FALSE
o percent não será alterado
como a resposta (answer_label) vai para o legenda do gráfico
quebra do answer_label em multiplas linhas com a função break_text com o parâmetro legend_max_size
caso a opção de ordenar as respostas por frequência, a resposta será ordenada pela percent
como a subquestion vai para o eixo x (que invertido fica como se fosse o eixo y)
quebra das subquestions coma função break_text utilizando como parâmetro axis_max_size
}
\author{
Leonardo Rocha
}
