% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{wilcox_ind_sample}
\alias{wilcox_ind_sample}
\title{Wilcox independent Sample}
\usage{
wilcox_ind_sample(data)
}
\arguments{
\item{data}{Tibble.}
}
\value{
row with Wilcoxon rank test for two indepedent columns
}
\description{
Utility function for the wilcoxon rank sum test
}
\author{
Leonardo Rocha
}
